pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mimic-triggerbench"
version = "0.0.1"
description = "MIMIC-TriggerBench: event-triggered ICU escalation benchmark and agent."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "MIMIC-TriggerBench Authors" }
]
dependencies = [
  "pydantic>=2.6",
  "pandas>=2.2",
  "pyarrow>=16.0",
  "duckdb>=1.0",
  "sqlalchemy>=2.0",
  "psycopg2-binary>=2.9",
  "scikit-learn>=1.4",
  "xgboost>=2.0",
  "python-dotenv>=1.0",
  "rich>=13.7",
  "pyyaml>=6.0",
]

[project.scripts]
mimic-triggerbench = "mimic_triggerbench.cli:main"

[project.optional-dependencies]
dev = [
  "pytest>=8.1",
  "pytest-cov>=5.0",
  "mypy>=1.9",
]

[tool.setuptools.packages.find]
where = ["src"]

